Fixed "too many open files" Error
Created an external file to look for online models to fix the bug causing the "Too many open files" error. 

added the ability to add the complete url for the model to the wanted.txt file, although I would still suggest adding just the models name as it appears in the url. 

few other small improvements.